feat: delete icon in recoil
